feat (subscription payment): Implement subscription payment.
Note:
According to the API doc, product and plan creations are one-time operations.
They should be created only if they have not been created before.
However, price in a plan is dynamic (based on the exchange rate).
Not sure if creating a new product and a new plan for every subscription (current implementation) will cause any issue.
If so, we will need to find a solution for this case.